Don't nest compute nodes and machine types
Currently the compute nodes and machine types are nested attributes like
this:

```hcl
nodes = {
  compute      = 10
  machine_type = "r5.xlarge"
}
```

But unfortunately this nesting produces an error that seems impossible
to solve when the `nodes` attribute isn't used, something like this:

```
Error: Value Conversion Error

  with ocm_cluster.my_cluster,
  on main.tf line 16, in resource "ocm_cluster" "my_cluster":
  16: resource "ocm_cluster" "my_cluster" {

An unexpected error was encountered trying to build a value. This is always
an error in the provider. Please report the following to the provider
developer:

unhandled null value
```

To avoid that this patch changes the attributes so that they aren't
nested:

```hcl
compute_nodes        = 10
compute_machine_type = "r5.xlarge"
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>