Use `cluster` instead of `cluster_id`
Currently the provider uses `cluster_id` in places where a reference to
a cluster is required. That isn't consistent with other places where
references are required. For example, references to cloud providers use
`cloud_provider` and references to machine types use `machine_type`. To
make that a bit more consistent this place replaces `cluster_id` with
`cluster`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>