Automatically prefix resources with the aws username
It has come to our attention that some of the resources are untagged.
This makes it hard to correlate them with the initiating user,
especially for cases where it's being run against a local CS env.

Here we first query the username from ``aws`` client (a preconfigured
``aws`` client is either way a requirement for using the provider) and
prefixes the resources with this value.