Refactoring the cluster properties validator
In the validator, we make sure that the user is not passing a
"reserved/default" property key. This was done by checking the key
against the 2 options: `rosa_tf_version` and `rosa_tf_commit`.

This refactor will be more future-proof. We can add more reserved
properties and not touch the validator in that area.

We now try to get the value of the proposed key from the OCMProperties
map. If found, we throw the error.

Signed-off-by: Sagi Dayan <sagidayan@gmail.com>