Add a PlanModifier to indicate an attribute is immutable.
This replaces the old ValueCannotBeChanged modifier that existed before
the framework upgrade. This modifier will only allow an attribute's
value to be set during the initial creation of the containing resource.
After that, any attempt to change the attribute will result in an error.

This modifier should be viewed as an alternative to the RequiresReplace
modifier where we would rather return an error to the user than cause
the resource to be recreated.

The modifier cannot be used on a "computed" attribute because computed
attributes may get assigned values that are not present in the
configuration.

It also does not permit "unknown" values (after the initial create) since
they cannot be validated.

Signed-off-by: John Strunk <jstrunk@redhat.com>