OCM-7317 | feat: add support for registry config
This PR adds support to configure the registries in HCP clusters.

The new feature is optional.

The input format is aligned with the one exposed by the APIs (which is
coming from OCP API themselves, to facilitate user adoption).

Validation in the provider is kept to a minimum initially, as it is done on backend
side.

Code is covered by unit testing and subsystem tests.

There is an additional Make target to generate code coverage from unit tests.

Example of input:
```
resource "rhcs_cluster_rosa_hcp" "rosa_hcp_cluster" {
  ...

  registry_config = {
    registry_sources = {
      allowed_registries = null
      blocked_registries = ["blocked.io", "blocked2.io"]
      insecure_registries = null
    }
    allowed_registries_for_import = [
    {
      domain_name = "registry.io"
      insecure = false
    }
    ]
  }
}
```