OCM-11554 | fix: empty list as default value for registry config
As the API is returning empty lists to clear parameters, while null values
are not used, we set empty list as default, so this will cover both null values
and empty lists in input, as they should behave in the same way.

Add also a custom validator for conflicts as we need to check against not empty list,while the default conflict validator checks only for null values.

Tests added and validated also locally the different cases.

Related: OCM-11554