OCM-17531 | feat: Enable refresh token support in OCM API auth flow
Before this commit, it was impossible to authenticate to OCM API endpoints which
require refresh tokens, such as the GovCloud internal OCM api. This made it
impossible to use the provider with ROSA GovCloud internally.

This commit adds support by:

- Exposing a new `refresh_token` provider variable which is appended to the
  OCM authentication builder token list

- Tweaks the client secret handling to unconditionally pass the client secret
  variable to the auth builder, even when empty. From the OCM SDK docs:

        // Note that some OpenID providers (Keycloak, for example) require the client identifier also for
        // the resource owner password grant. In that case use the set only the identifier, and let the
        // secret blank. For example:

  This change should thus support all workflows in a backwards compatible way.

After this change, the provider can successfully authenticate to ROSA GovCloud
OCM API endpoints when appropriately configured.