WIP: Make billing account ID optional to enable GovCloud HCP cluster creation
Before this commit, HCP clusters on ROSA GovCloud couldn't be created because the
module required a value for the billing account variable, which is invalid to
supply to the OCM API when creating HCP GovCloud clusters[1]. In these cases,
the billing ID passed to the OCM `/clusters` API should be empty.

This commit makes billing account ID optional so that users can create HCP clusters
on ROSA GovCloud.

[1] https://github.com/openshift/rosa/blob/master/cmd/create/cluster/cmd.go#L1268